Cleanup and Compiler Warnings (#6)

+++ src/logging.cc
@@ -10,7 +10,9 @@ scoped_timer::scoped_timer(std::string name)
 scoped_timer::~scoped_timer() {
   using namespace std::chrono;
   auto stop = steady_clock::now();
-  double t = duration_cast<microseconds>(stop - start_).count() / 1000.0;
+  double t =
+      static_cast<double>(duration_cast<microseconds>(stop - start_).count()) /
+      1000.0;
   log(log_lvl::info, name_.c_str(), "finished {}ms", t);
 }
